refactor: Use Class instances instead of MetaClass for mapper objects
