+++ .env.development
@@ -1,4 +1,4 @@
-AHD_API_TOKEN=d13afc8caabcddd48a91448dc357cfab1b823641bdf0befa15506e96eedae676
+AHD_API_TOKEN=7270dedcac3aeb0a46ff302ae11f3eec4041ad03fe11f96c7400a4dc470c9061
 AHD_PIPELINE=discharge
 AHD_PROJECT=test
 AHD_URL=http://localhost:9999/health-discovery
