+++ ahd2fhir/utils/resource_handler.py
@@ -1,28 +1,17 @@
-import base64
-import datetime
 import logging
-import time
-from typing import List, Tuple
+from typing import List
 
 import structlog
 import tenacity
 from averbis import Pipeline
 from fhir.resources.bundle import Bundle
 from fhir.resources.codeableconcept import CodeableConcept
-from fhir.resources.composition import Composition, CompositionSection
 from fhir.resources.documentreference import DocumentReference
-from fhir.resources.fhirtypes import DateTime
-from fhir.resources.identifier import Identifier
-from fhir.resources.reference import Reference
-from fhir.resources.resource import Resource
 from prometheus_client import Counter, Histogram, Summary
 from tenacity.after import after_log
 
-# from ahd2fhir.mappers import ahd_to_condition, ahd_to_medication_statement
 from ahd2fhir.utils.bundle_builder import BundleBuilder
-
-# from ahd2fhir.utils.device_builder import build_device
-from ahd2fhir.utils.fhir_utils import sha256_of_identifier
+from ahd2fhir.utils.fhir_utils import _build_composition, _extract_text_from_resource
 
 MAPPING_FAILURES_COUNTER = Counter("mapping_failures", "Exceptions during mapping")
 MAPPING_DURATION_SUMMARY = Histogram(
@@ -97,7 +86,7 @@ class ResourceHandler:
             resources_from_document = self._process_documentreference(
                 document_reference
             )
-            composition = self._build_composition(
+            composition = _build_composition(
                 document_reference, resources_from_document
             )
 
@@ -125,89 +114,6 @@ class ResourceHandler:
 
         return self.handle_documents(document_references)
 
-    def _build_composition(
-        self, document_reference: DocumentReference, all_resources: List[Resource]
-    ):
-        composition_type = (
-            document_reference.type
-            if document_reference.type is not None
-            else DISCHARGE_SUMMARY_CONCEPT
-        )
-
-        composition_subject = document_reference.subject
-        composition_category = document_reference.category
-        composition_encounter = None
-
-        if document_reference.context is not None:
-            if len(document_reference.context.encounter) > 1:
-                log.warning(
-                    "DocumentReference contains more than one encounter. "
-                    + "Using the first."
-                )
-            composition_encounter = document_reference.context.encounter[0]
-
-        composition_author = None
-        composition_sections = []
-        for resource in all_resources:
-            resource_type = resource.resource_type
-
-            if resource_type == "Device":
-                author = Reference.construct()
-                author.reference = f"Device/{resource.id}"
-                author.type = "Device"
-                composition_author = author
-                continue
-
-            # Check if no resource specific section exists ands adds it,
-            # otherwise select the correct section
-            if not any(
-                section.title == resource_type for section in composition_sections
-            ):
-                resource_section = CompositionSection.construct()
-                resource_section.title = resource_type
-                resource_section.entry = []
-
-                composition_sections.append(resource_section)
-
-                ind = len(composition_sections) - 1
-            else:
-                ind = [
-                    ind
-                    for ind, section in enumerate(composition_sections)
-                    if section.title == resource_type
-                ][0]
-
-            entry_reference = Reference.construct()
-            entry_reference.reference = resource_type + "/" + resource.id
-
-            composition_sections[ind].entry.append(entry_reference)
-
-        if composition_author is None:
-            composition_author = Reference(**{"display": "Averbis Health Discovery"})
-
-        composition_identifier = (
-            self._build_composition_identifier_from_documentreference(
-                document_reference
-            )
-        )
-
-        composition = Composition(
-            **{
-                "title": "NLP FHIR Results " + time.strftime("%Y-%m-%dT%H:%M"),
-                "status": "final",
-                "date": DateTime.validate(datetime.datetime.now(datetime.timezone.utc)),
-                "type": composition_type,
-                "identifier": composition_identifier,
-                "id": sha256_of_identifier(composition_identifier),
-                "subject": composition_subject,
-                "category": composition_category,
-                "encounter": composition_encounter,
-                "author": [composition_author],
-                "section": composition_sections,
-            }
-        )
-        return composition
-
     def _process_documentreference(self, document_reference: DocumentReference):
         log = structlog.get_logger().bind(
             document_id=f"{document_reference.get_resource_type()}/"
@@ -215,9 +121,7 @@ class ResourceHandler:
         )
 
         # Text extraction and text analysis
-        (text, content_type, lang) = self._extract_text_from_resource(
-            document_reference
-        )
+        (text, content_type, lang) = _extract_text_from_resource(document_reference)
 
         DOCUMENT_LENGTH_SUMMARY.observe(len(text))
 
@@ -270,38 +174,6 @@ class ResourceHandler:
 
         return total_results
 
-    def _extract_text_from_resource(
-        self,
-        document_reference: DocumentReference,
-    ) -> Tuple[str, str]:
-        valid_content = [
-            content
-            for content in document_reference.content
-            if content.attachment.data is not None
-        ]
-
-        if len(valid_content) == 0:
-            raise ValueError(
-                f"Document {document_reference.id} contains no valid content"
-            )
-
-        if len(valid_content) > 1:
-            raise ValueError(
-                f"Document {document_reference.id} contains more than one attachment"
-            )
-
-        content = valid_content[0]
-
-        language = None
-        if content.attachment.language:
-            language = content.attachment.language.lower().split("-")[0]
-
-        return (
-            base64.b64decode(content.attachment.data).decode("utf8"),
-            content.attachment.contentType,
-            language,
-        )
-
     @tenacity.retry(
         stop=tenacity.stop.stop_after_attempt(10),
         wait=tenacity.wait.wait_fixed(5)
@@ -312,15 +184,7 @@ class ResourceHandler:
     def _perform_text_analysis(
         self, text: str, mime_type: str = "text/plain", lang: str = None
     ):
-        types = ",".join(
-            [
-                # AHD_TYPE_DIAGNOSIS,
-                # AHD_TYPE_MEDICATION,
-                # AHD_TYPE_DOCUMENT_ANNOTATION,
-                *self.mapper_handler.get_ahd_types(),
-            ]
-        )
-
+        types = ",".join([*self.mapper_handler.get_ahd_types()])
         analyse_args = {"language": lang, "annotation_types": types}
 
         try:
@@ -332,40 +196,3 @@ class ResourceHandler:
             log.exception(exc)
             log.error("Text analysis failed")
             raise exc
-
-    def _build_composition_identifier_from_documentreference(
-        self,
-        doc_ref: DocumentReference,
-    ):
-        """
-        construct a hopefully unqiue identifier for the condition from
-        the document identifier as well as the offset into the text
-        and the unique id of the annotation
-        """
-        doc_ref_identifier = None
-        if doc_ref.identifier is None or len(doc_ref.identifier) == 0:
-            log.warning(
-                "No identifier specified on the document. "
-                + "Trying to fall-back to the DocumentReference.id"
-            )
-            doc_ref_identifier = doc_ref.id
-        else:
-            if len(doc_ref.identifier) > 1:
-                log.warning(
-                    "More than one identifier specified on the document. "
-                    + "Using the first occurrence."
-                )
-            doc_ref_identifier = doc_ref.identifier[0].value
-
-        composition_identifier_system = (
-            "https://fhir.miracum.org/nlp/identifiers/ahd-analysis-result-composition"
-        )
-
-        composition_identifier_value = f"{doc_ref_identifier}_ahd-analysis-result"
-
-        return Identifier(
-            **{
-                "system": composition_identifier_system,
-                "value": composition_identifier_value,
-            }
-        )
